專題-NTUSTeam_linebot/line_chatbot/requirements.txt
line-bot-sdk==1.8.0
flask==1.0.2
gunicorn
requests==2.21.0
beautifulsoup4==4.7.1
selenium==3.8.0
google-api-python-client
oauth2client